Websites/Face_Detector/requirements.txt
Flask==2.1.2
Flask_SocketIO==5.2.0
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.2.0
pip==22.1.2